requirements.txt
flask
flask_mail
python-dotenv
notion-client
markdown2
markdown2